refactor!: overhaul architecture from demo to production-ready standards
Major architectural restructuring to address security, performance, and maintainability issues flagged in code review.

Key Changes:

🔒 Security & Configuration:
- Removed all hardcoded secrets (JWT, DB passwords) from source code.
- Integrated Viper for configuration management with Environment Variable override support.
- Masked internal error details in HTTP 500 responses to prevent info disclosure.
- Hardened Docker Compose with resource limits (CPU/Memory) and non-root considerations.

🏗 Architecture & Logic:
- Refactored Product API: `Attributes` is now a JSON object (map), not a string.
- Implemented atomic stock deduction in Repository layer to ensure concurrency safety.
- Fixed nondeterministic pagination by enforcing `ORDER BY` in SQL queries.
- Added explicit input validation for negative limits/offsets.

⚙️ DevOps & Tooling:
- Optimized Makefile: Replaced hardcoded mock lists with `go generate ./...`.
- Enhanced CI/CD: Added `golangci-lint` and `-race` detection in GitHub Actions.
- Fixed test suite: Removed `gomock.Any()` usage for strict parameter assertion.
- Added extensive Unit Tests for Order, Product, and User modules.

BREAKING CHANGE:
- The `CreateProduct` API payload structure for `attributes` has changed from String to JSON Object.
- Configuration now requires specific Environment Variables (e.g., MALL_JWT_SECRET) to run.

+++ internal/mocks/product_service_mock.go
@@ -57,7 +57,7 @@ func (mr *MockProductServiceMockRecorder) CreateProduct(ctx, req any) *gomock.Ca
 }
 
 // GetProduct mocks base method.
-func (m *MockProductService) GetProduct(ctx context.Context, spuID uint) (*service.ProductResp, error) {
+func (m *MockProductService) GetProduct(ctx context.Context, spuID uint64) (*service.ProductResp, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetProduct", ctx, spuID)
 	ret0, _ := ret[0].(*service.ProductResp)
